Refactoring into go file layout
