+++ cmd/bitwardend/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/VictorNine/bitwarden-go/pkg/server"
+
+func main() {
+	server.Start()
+}
