+++ pkg/server/database_mock.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import _ "github.com/mattn/go-sqlite3"
